Renamed Hazel Editor to Hazelnut, created viewport panel for scene
