--- Hazel-Editor/imgui.ini
@@ -1,19 +1,0 @@
-[Window][DockSpace Demo]
-Pos=0,0
-Size=1280,720
-Collapsed=0
-
-[Window][Debug##Default]
-Pos=60,60
-Size=400,400
-Collapsed=0
-
-[Window][Settings]
-Pos=0,19
-Size=1280,701
-Collapsed=0
-DockId=0x3BC79352,0
-
-[Docking][Data]
-DockSpace ID=0x3BC79352 Window=0x4647B76E Pos=77,227 Size=1280,701 CentralNode=1 Selected=0x1C33C293
-
