+++ Hazel/src/Hazel/Renderer/Framebuffer.h
@@ -19,6 +19,8 @@ namespace Hazel {
 		virtual void Bind() = 0;
 		virtual void Unbind() = 0;
 
+		virtual void Resize(uint32_t width, uint32_t height) = 0;
+
 		virtual uint32_t GetColorAttachmentRendererID() const = 0;
 
 		virtual const FramebufferSpecification& GetSpecification() const = 0;
