+++ Hazel/src/Hazel/Renderer/OrthographicCameraController.h
@@ -16,6 +16,8 @@ namespace Hazel {
 		void OnUpdate(Timestep ts);
 		void OnEvent(Event& e);
 
+		void OnResize(float width, float height);
+
 		OrthographicCamera& GetCamera() { return m_Camera; }
 		const OrthographicCamera& GetCamera() const { return m_Camera; }
