+++ Hazel/src/Platform/OpenGL/OpenGLFramebuffer.h
@@ -15,12 +15,14 @@ namespace Hazel {
 		virtual void Bind() override;
 		virtual void Unbind() override;
 
+		virtual void Resize(uint32_t width, uint32_t height) override;
+
 		virtual uint32_t GetColorAttachmentRendererID() const override { return m_ColorAttachment; }
 
 		virtual const FramebufferSpecification& GetSpecification() const override { return m_Specification; }
 	private:
-		uint32_t m_RendererID;
-		uint32_t m_ColorAttachment, m_DepthAttachment;
+		uint32_t m_RendererID = 0;
+		uint32_t m_ColorAttachment = 0, m_DepthAttachment = 0;
 		FramebufferSpecification m_Specification;
 	};
