+++ Hazelnut/assets/shaders/FlatColor.glsl
