+++ Hazelnut/imgui.ini
@@ -1,0 +1,28 @@
+[Window][DockSpace Demo]
+Pos=0,0
+Size=1933,431
+Collapsed=0
+
+[Window][Debug##Default]
+ViewportPos=397,284
+ViewportId=0x9F5F46A1
+Size=400,400
+Collapsed=0
+
+[Window][Settings]
+Pos=0,19
+Size=339,412
+Collapsed=0
+DockId=0x00000001,0
+
+[Window][Viewport]
+Pos=341,19
+Size=1592,412
+Collapsed=0
+DockId=0x00000002,0
+
+[Docking][Data]
+DockSpace   ID=0x3BC79352 Window=0x4647B76E Pos=580,541 Size=1933,412 Split=X Selected=0x995B0CF8
+  DockNode  ID=0x00000001 Parent=0x3BC79352 SizeRef=339,701 Selected=0x1C33C293
+  DockNode  ID=0x00000002 Parent=0x3BC79352 SizeRef=2219,701 CentralNode=1 Selected=0x995B0CF8
+
