+++ Hazelnut/src/EditorLayer.h
@@ -26,6 +26,9 @@ namespace Hazel {
 
 		Ref<Texture2D> m_CheckerboardTexture;
 
+		glm::vec2 m_ViewportSize = { 0.0f, 0.0f };
+		
+
 		glm::vec4 m_SquareColor = { 0.2f, 0.3f, 0.8f, 1.0f };
 	};
