+++ Hazelnut/src/HazelnutApp.cpp
@@ -5,23 +5,23 @@
 
 namespace Hazel {
 
-	class HazelEditor : public Application
+	class Hazelnut : public Application
 	{
 	public:
-		HazelEditor()
-			: Application("Hazel Editor")
+		Hazelnut()
+			: Application("Hazelnut")
 		{
 			PushLayer(new EditorLayer());
 		}
 
-		~HazelEditor()
+		~Hazelnut()
 		{
 		}
 	};
 
 	Application* CreateApplication()
 	{
-		return new HazelEditor();
+		return new Hazelnut();
 	}
 
 }
